split introspection into multiple files implement source fields and enum values filtering properly configure field mapping for introspection fields
